divide index page into multiple components.
