+++ app/components/company-links.js
@@ -1,0 +1,6 @@
+import Ember from 'ember';
+
+export default Ember.Component.extend({
+  classNames: ['nav'],
+  tagName: 'ul'
+});
