+++ app/templates/components/company-links.hbs
@@ -1,0 +1,5 @@
+<li><a href="index.html">Home</a></li>
+<li><a href="blog-right-sidebar.html">Blog</a></li>
+<li><a href="portfolio-3col.html">Portfolio</a></li>
+<li><a href="page-about.html">About</a></li>
+<li><a href="page-contact.html">Contact</a></li>
