+++ app/templates/components/copy-right.hbs
@@ -1,0 +1,2 @@
+<p>Copyright © 2015 HedgeServ by <a target="_blank" href="http://www.hedgeserv.com">HedgeServ</a>. All Rights
+    Reserved</p>
