+++ app/templates/components/social-links.hbs
@@ -1,0 +1,16 @@
+<li class="facebook"><a target="_blank" href="http://www.facebook.com"><i
+        class="fa fa-facebook"></i></a></li>
+<li class="twitter"><a target="_blank" href="http://www.twitter.com"><i
+        class="fa fa-twitter"></i></a></li>
+<li class="googleplus"><a target="_blank" href="http://plus.google.com"><i
+        class="fa fa-google-plus"></i></a></li>
+<li class="skype"><a target="_blank" href="http://www.skype.com"><i class="fa fa-skype"></i></a>
+</li>
+<li class="linkedin"><a target="_blank" href="http://www.linkedin.com"><i
+        class="fa fa-linkedin"></i></a></li>
+<li class="youtube"><a target="_blank" href="http://www.youtube.com"><i
+        class="fa fa-youtube-play"></i></a></li>
+<li class="flickr"><a target="_blank" href="http://www.flickr.com"><i
+        class="fa fa-flickr"></i></a></li>
+<li class="pinterest"><a target="_blank" href="http://www.pinterest.com"><i
+        class="fa fa-pinterest"></i></a></li>
