+++ app/templates/components/subscribe-form.hbs
@@ -1,0 +1,9 @@
+<form>
+    <div class="form-group has-feedback">
+        <label class="sr-only" for="subscribe">Email address</label>
+        <input type="email" class="form-control" id="subscribe" placeholder="Enter email"
+               name="subscribe" required>
+        <i class="fa fa-envelope form-control-feedback"></i>
+        <button type="submit" class="btn btn-white btn-sm">Submit</button>
+    </div>
+</form>
