+++ app/templates/components/testimonial-show.hbs
@@ -1,0 +1,64 @@
+<div class="testimonial">
+    <div class="container">
+        <div class="row">
+            <div class="col-md-8 col-md-offset-2">
+                <h2 class="title">Just Perfect!</h2>
+
+                <div class="testimonial-image">
+                    <img src="images/testimonial-1.jpg" alt="Jane Doe" title="Jane Doe" class="img-circle">
+                </div>
+                <div class="testimonial-body">
+                    <p>Sed ut perspiciatis unde omnis iste natu error sit voluptatem accusan tium dolore laud
+                        antium, totam rem dolor sit amet tristique pulvinar, turpis arcu rutrum nunc, ac laoreet
+                        turpis augue a justo.</p>
+
+                    <div class="testimonial-info-1">- Jane Doe</div>
+                    <div class="testimonial-info-2">By Company</div>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+<div class="testimonial">
+    <div class="container">
+        <div class="row">
+            <div class="col-md-8 col-md-offset-2">
+                <h2 class="title">Wow amazing!</h2>
+
+                <div class="testimonial-image">
+                    <img src="images/testimonial-2.jpg" alt="Jane Doe" title="Jane Doe" class="img-circle">
+                </div>
+                <div class="testimonial-body">
+                    <p>Turpis arcu rutrum nunc, ac laoreet turpis augue a justo. Lorem ipsum dolor sit amet,
+                        consectetur adipisicing elit. Expedita, esse, animi maxime voluptate tempore at sunt
+                        labore incidunt nulla dignissimos iusto ad similique inventore distinctio quam
+                        repellendus itaque minus minima.</p>
+
+                    <div class="testimonial-info-1">- John Doe</div>
+                    <div class="testimonial-info-2">By Company</div>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+<div class="testimonial">
+    <div class="container">
+        <div class="row">
+            <div class="col-md-8 col-md-offset-2">
+                <h2 class="title">Thanks!</h2>
+
+                <div class="testimonial-image">
+                    <img src="images/testimonial-3.jpg" alt="Jane Doe" title="Jane Doe" class="img-circle">
+                </div>
+                <div class="testimonial-body">
+                    <p>Phosfluorescently e-enable adaptive synergy for strategic quality vectors. Continually
+                        transform fully tested expertise with competitive technologies ac laoreet turpis augue a
+                        justo.</p>
+
+                    <div class="testimonial-info-1">- John Doe</div>
+                    <div class="testimonial-info-2">By Company</div>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
