+++ app/templates/components/who-we-are.hbs
@@ -1,0 +1,89 @@
+
+<h1 class="text-center">Who we are</h1>
+<div class="separator"></div>
+<p class="lead text-center">Lorem ipsum dolor sit amet laudantium molestias similique.<br> Quisquam
+    incidunt ut laboriosam.</p>
+<br>
+<div class="row">
+    <div class="col-md-6">
+        <h2 class="title">iDea - Responsive Website Theme</h2>
+        <div class="row">
+            <div class="col-md-6">
+                <img src="images/section-image-2.png" alt="">
+            </div>
+            <div class="col-md-6">
+                <p>Quo soluta provident, quod reiciendis. Dolores nam totam aut illum ex ratione harum
+                    molestias maxime minima tempore, possimus, laudantium. Lorem ipsum dolor sit amet,
+                    consectetur adipisicing elit.</p>
+            </div>
+        </div>
+        <p>Esse sequi veniam, assumenda voluptate necessitatibus ipsa dicta vero, minima natus cum
+            cupiditate magnam et placeat quo adipisci.</p>
+        <a href="page-about.html" class="btn btn-white">Learn More</a>
+        <div class="space hidden-md hidden-lg"></div>
+    </div>
+    <div class="col-md-6">
+        <div class="panel-group panel-dark" id="accordion">
+            <div class="panel panel-default">
+                <div class="panel-heading">
+                    <h4 class="panel-title">
+                        <a data-toggle="collapse" data-parent="#accordion" href="#collapseOne">
+                            <i class="fa fa-bold"></i>Bootstrap Based
+                        </a>
+                    </h4>
+                </div>
+                <div id="collapseOne" class="panel-collapse collapse in">
+                    <div class="panel-body">
+                        Anim pariatur cliche reprehenderit, enim eiusmod high life accusamus terry
+                        richardson ad squid. 3 wolf moon officia aute, non cupidatat skateboard dolor
+                        brunch. Food truck quinoa nesciunt laborum eiusmod. Brunch 3 wolf moon tempor,
+                        sunt aliqua put a bird on it squid single-origin coffee nulla assumenda
+                        shoreditch et. Nihil anim keffiyeh helvetica, craft beer labore wes anderson
+                        cred nesciunt sapiente ea proident.
+                    </div>
+                </div>
+            </div>
+            <div class="panel panel-default">
+                <div class="panel-heading">
+                    <h4 class="panel-title">
+                        <a data-toggle="collapse" data-parent="#accordion" href="#collapseTwo"
+                           class="collapsed">
+                            <i class="fa fa-leaf"></i>Clean &amp; powerful code
+                        </a>
+                    </h4>
+                </div>
+                <div id="collapseTwo" class="panel-collapse collapse">
+                    <div class="panel-body">
+                        Anim pariatur cliche reprehenderit, enim eiusmod high life accusamus terry
+                        richardson ad squid. 3 wolf moon officia aute, non cupidatat skateboard dolor
+                        brunch. Food truck quinoa nesciunt laborum eiusmod. Brunch 3 wolf moon tempor,
+                        sunt aliqua put a bird on it squid single-origin coffee nulla assumenda
+                        shoreditch et. Nihil anim keffiyeh helvetica, craft beer labore wes anderson
+                        cred nesciunt sapiente ea proident.
+                    </div>
+                </div>
+            </div>
+            <div class="panel panel-default">
+                <div class="panel-heading">
+                    <h4 class="panel-title">
+                        <a data-toggle="collapse" data-parent="#accordion" href="#collapseThree"
+                           class="collapsed">
+                            <i class="fa fa-html5"></i>HTML5/CSS3
+                        </a>
+                    </h4>
+                </div>
+                <div id="collapseThree" class="panel-collapse collapse">
+                    <div class="panel-body">
+                        Anim pariatur cliche reprehenderit, enim eiusmod high life accusamus terry
+                        richardson ad squid. 3 wolf moon officia aute, non cupidatat skateboard dolor
+                        brunch. Food truck quinoa nesciunt laborum eiusmod. Brunch 3 wolf moon tempor,
+                        sunt aliqua put a bird on it squid single-origin coffee nulla assumenda
+                        shoreditch et. Nihil anim keffiyeh helvetica, craft beer labore wes anderson
+                        cred nesciunt sapiente ea proident.
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+<br>
